pyproject.toml
[tool.poetry]
name = "metronome"
version = "0.1.0"
description = ""
authors = ["Christian Weber"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
